# Always inline and remove use of StepRange