--- a/src/rte/longwave1scalar.jl
+++ b/src/rte/longwave1scalar.jl
@@ -39,7 +39,7 @@ function rte_lw_noscat_solve!(
     tx = min(ncol, max_threads)
     bx = cld(ncol, tx)
     args = (flux_lw, src_lw, bcs_lw, op, nlay, ncol, as)
-    @cuda threads = (tx) blocks = (bx) rte_lw_noscat_solve_CUDA!(args...)
+    @cuda always_inline = true threads = (tx) blocks = (bx) rte_lw_noscat_solve_CUDA!(args...)
     return nothing
 end
 
@@ -125,7 +125,7 @@ function rte_lw_noscat_solve!(
     tx = min(ncol, max_threads)
     bx = cld(ncol, tx)
     args = (flux, flux_lw, src_lw, bcs_lw, op, nlay, ncol, as, lookup_lw, lookup_lw_cld)
-    @cuda threads = (tx) blocks = (bx) rte_lw_noscat_solve_CUDA!(args...)
+    @cuda always_inline = true threads = (tx) blocks = (bx) rte_lw_noscat_solve_CUDA!(args...)
     return nothing
 end