--- a/src/rte/longwave1scalar.jl
+++ b/src/rte/longwave1scalar.jl
@@ -244,7 +244,8 @@ Transport for no-scattering longwave problem.
 
     # Top of domain is index nlev
     # Downward propagation
-    @inbounds for ilev in nlay:-1:1
+    ilev = nlay
+    @inbounds while ilev ≥ 1
         τ_loc = τ[ilev, gcol] * Ds[1]
         trans = exp(-τ_loc)
         lay_src, lev_src_dec = lay_source[ilev, gcol], lev_source_dec[ilev, gcol]
@@ -252,6 +253,7 @@ Transport for no-scattering longwave problem.
             trans * intensity_dn_ilevplus1 + lw_noscat_source_dn(lev_src_dec, lay_src, τ_loc, trans, τ_thresh)
         intensity_dn_ilevplus1 = intensity_dn_ilev
         flux_dn[ilev, gcol] = intensity_dn_ilev * intensity_to_flux
+        ilev -= 1
     end
 
     # Surface reflection and emission