--- a/src/rte/shortwave1scalar.jl
+++ b/src/rte/shortwave1scalar.jl
@@ -160,8 +160,10 @@ function rte_sw_noscat!(
     (; flux_dn_dir, flux_net) = flux
     # downward propagation
     @inbounds flux_dn_dir[nlev, gcol] = toa_flux[gcol] * solar_frac * cos_zenith[gcol]
-    @inbounds for ilev in (nlev - 1):-1:1
+    ilev = nlev - 1
+    @inbounds while ilev ≥ 1
         flux_dn_dir[ilev, gcol] = flux_dn_dir[ilev + 1, gcol] * exp(-τ[ilev, gcol] / cos_zenith[gcol])
         flux_net[ilev, gcol] = -flux_dn_dir[ilev, gcol]
+        ilev -= 1
     end
 end