--- a/src/rte/shortwave2stream.jl
+++ b/src/rte/shortwave2stream.jl
@@ -43,7 +43,7 @@ function rte_sw_2stream_solve!(
     tx = min(ncol, max_threads)
     bx = cld(ncol, tx)
     args = (flux_sw, op, bcs_sw, src_sw, nlay, ncol, as)
-    @cuda threads = (tx) blocks = (bx) rte_sw_2stream_solve_CUDA!(args...)
+    @cuda always_inline = true threads = (tx) blocks = (bx) rte_sw_2stream_solve_CUDA!(args...)
     return nothing
 end
 
@@ -151,7 +151,7 @@ function rte_sw_2stream_solve!(
     tx = min(ncol, max_threads)
     bx = cld(ncol, tx)
     args = (flux, flux_sw, op, bcs_sw, src_sw, nlay, ncol, as, lookup_sw, lookup_sw_cld)
-    @cuda threads = (tx) blocks = (bx) rte_sw_2stream_solve_CUDA!(args...)
+    @cuda always_inline = true threads = (tx) blocks = (bx) rte_sw_2stream_solve_CUDA!(args...)
     return nothing
 end