This PR:

 - applies `always_inline = true` for the cuda launches
 - Removes use of `StepRange`

as suggested in https://cuda.juliagpu.org/dev/tutorials/performance/