@sriharshakandala 2024-02-06
@charleskawczynski : Just curious how much this improves the times for the DYAMOND configuration! https://github.com/CliMA/RRTMGP.jl/pull/439