@charleskawczynski 2024-02-06
> @charleskawczynski : Just curious how much this improves the times for the DYAMOND configuration! #439

For the rrtmgp callback on p100 in the dyamond simulation:
 - main branch: ~26 seconds
 - this branch: ~19 seconds

cc @sriharshakandala