@sriharshakandala 2024-02-06
Current RRTMGP main on A100:
```
julia --project=gpuenv test/all_sky_tuning.jl 
device = ClimaComms.CUDADevice(); FT = Float64, ncols = 131658; size per field = 0.04119899868965149 GB
"timing longwave solver" = "timing longwave solver"
  1.251112 seconds (98 CPU allocations: 17.938 KiB)
  1.251308 seconds (97 CPU allocations: 17.859 KiB)
  1.250872 seconds (98 CPU allocations: 17.969 KiB)
  1.250838 seconds (77 CPU allocations: 16.062 KiB)
  1.250489 seconds (77 CPU allocations: 16.062 KiB)
"timing shortwave solver" = "timing shortwave solver"
  0.930035 seconds (77 CPU allocations: 16.125 KiB)
  0.930778 seconds (77 CPU allocations: 16.125 KiB)
  0.930711 seconds (77 CPU allocations: 16.125 KiB)
  0.930512 seconds (77 CPU allocations: 16.125 KiB)
  0.930794 seconds (77 CPU allocations: 16.125 KiB)
 39.934982 seconds (98.69 M allocations: 5.615 GiB, 4.45% gc time, 53.17% compilation time: <1% of which was recompilation)
```
This PR
```
julia --project=gpuenv test/all_sky_tuning.jl 
device = ClimaComms.CUDADevice(); FT = Float64, ncols = 131658; size per field = 0.04119899868965149 GB
"timing longwave solver" = "timing longwave solver"
  1.218868 seconds (98 CPU allocations: 17.938 KiB)
  1.219475 seconds (97 CPU allocations: 17.859 KiB)
  1.219142 seconds (98 CPU allocations: 17.969 KiB)
  1.220116 seconds (77 CPU allocations: 16.062 KiB)
  1.219724 seconds (77 CPU allocations: 16.062 KiB)
"timing shortwave solver" = "timing shortwave solver"
  0.884353 seconds (77 CPU allocations: 16.125 KiB)
  0.884920 seconds (77 CPU allocations: 16.125 KiB)
  0.884993 seconds (77 CPU allocations: 16.125 KiB)
  0.884079 seconds (77 CPU allocations: 16.125 KiB)
  0.884311 seconds (77 CPU allocations: 16.125 KiB)
 38.858708 seconds (98.92 M allocations: 5.617 GiB, 4.01% gc time, 53.04% compilation time: 1% of which was recompilation)
```